factor out pattern templates
Also fix incorrect test splitting marker
