+++ src/enzyme_ad/jax/Passes/SelectPatterns.h
@@ -1,0 +1,17 @@
+//===- SelectPatterns.h - Patterns to simplify 'select' ops --*- C++ -*----===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+namespace mlir {
+class RewritePatternSet;
+
+namespace enzyme {
+
+void populateSelectExtractPatterns(RewritePatternSet &patterns);
+
+} // namespace enzyme
+} // namespace mlir
